create DescribedMixin and start refactoring of utils
