+++ common/models/mixins/description.py
@@ -5,6 +5,7 @@ from polymorphic.managers import PolymorphicManager
 
 from common.business_rules import NoBlankDescription
 from common.business_rules import UpdateValidity
+from common.exceptions import NoDescriptionError
 from common.models.mixins.validity import ValidityStartMixin
 from common.models.mixins.validity import ValidityStartQueryset
 from common.models.records import TrackedModelQuerySet
@@ -62,3 +63,63 @@ class DescriptionMixin(ValidityStartMixin):
 
     class Meta:
         abstract = True
+
+
+class DescribedMixin:
+    """Mixin adding convenience methods for TrackedModels with associated
+    Descriptions."""
+
+    def get_descriptions(self, transaction=None, request=None) -> TrackedModelQuerySet:
+        """
+        Get the latest descriptions related to this instance of the Tracked
+        Model.
+
+        If there is no Description relation existing a `NoDescriptionError` is raised.
+
+        If a transaction is provided then all latest descriptions that are either approved
+        or in the workbasket of the transaction up to the transaction will be provided.
+        """
+        try:
+            descriptions_model = self.descriptions.model
+        except AttributeError as e:
+            raise NoDescriptionError(
+                f"Model {self.__class__.__name__} has no descriptions relation.",
+            ) from e
+
+        for field, model in descriptions_model.relations.items():
+            if isinstance(self, model):
+                field_name = field.name
+                break
+        else:
+            raise NoDescriptionError(
+                f"No foreign key back to model {self.__class__.__name__} "
+                f"found on description model {descriptions_model.__name__}.",
+            )
+
+        filter_kwargs = {
+            f"{field_name}__{key}": value
+            for key, value in self.get_identifying_fields().items()
+        }
+
+        query = descriptions_model.objects.filter(**filter_kwargs).order_by(
+            *descriptions_model._meta.ordering
+        )
+
+        if transaction:
+            return query.approved_up_to_transaction(transaction=transaction)
+
+        if request:
+            from workbaskets.models import WorkBasket
+
+            return query.approved_up_to_transaction(
+                transaction=WorkBasket.get_current_transaction(request),
+            )
+
+        return query.latest_approved()
+
+    def get_description(self):
+        return self.get_descriptions(transaction=self.transaction).last()
+
+    @property
+    def structure_description(self) -> str:
+        return self.get_description().description
